Урок 2 усправление

+++ 02_lesson/lesson_2_task_3.py
@@ -3,5 +3,5 @@ import math
 def square(side):
     return math.ceil(side*side)
 
-num_side = int(input("Сторона квадрата: "))
+num_side = float(input("Сторона квадрата: "))
 print(f"Площадь квадрата: {square(num_side)}")
